+++ CHANGES.md
@@ -14,10 +14,13 @@ Change Log
     * Removed `HeightmapTessellator` from the public API. It is now private and subject to change without notice.
     * Removed `TerrainMesh` from the public API. It is now private and subject to change without notice.
     * Removed `jsonp`. Use `loadJsonp` instead.
+* Deprecated
+    * Deprecated `WebMapTileServiceImageryProvider`. It will be removed in 1.19. Use `createWebMapTileServiceImageryProvider` instead.
 * Reduced the amount of both GPU and CPU memory used by terrain. The CPU memory was reduced by up to 40%.
 * `CorridorGeometry` and `PolylineVolumeGeometry` render short segments [#3293](https://github.com/AnalyticalGraphicsInc/cesium/issues/3293)
 * `Rectangle.fromCartographicArray` finds the smallest rectangle regardess of whether or not it crosses the international date line. [#3227](https://github.com/AnalyticalGraphicsInc/cesium/issues/3227)
 * Bug fix for `CorridorGeometry` with nearly colinear points [#3320](https://github.com/AnalyticalGraphicsInc/cesium/issues/3320)
+* Added `createWebMapTileServiceImageryProvider.js` function to replace the `WebMapTileServiceImageryProvider.js` class. This function returns a constructed `UrlTemplateImageryProvider`.
 
 ### 1.16 - 2015-12-01